services/gps_nats/requirements.txt
postgres
gpsd-py3
prometheus-client
asyncio-nats-client